get_address: update for stax
